Avoid -fsanitize=signed-integer-overflow
...as seen with `--convert-to epub doc/ooo111691-2.doc` (i.e., attachment
111691_a.doc at <https://bz.apache.org/ooo/show_bug.cgi?id=111691#c3>)

+++ include/svx/sdasitm.hxx
@@ -88,7 +88,7 @@ inline SdrOnOffItem makeSdrTextWordWrapItem( bool bAuto ) {
 
 inline size_t SdrCustomShapeGeometryItem::PropertyPairHash::operator()( const SdrCustomShapeGeometryItem::PropertyPair &r1 ) const
 {
-    return r1.first.hashCode() * 37 + r1.second.hashCode();
+    return size_t(r1.first.hashCode()) * 37 + size_t(r1.second.hashCode());
 };
 
 #endif
